requirements.txt
pydantic~=2.11.4
python-dotenv
